Move models from Exchanges api entry to models dedicated package
